requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.4
django-cache-url==3.4.5
django-ckeditor-5==0.2.12
django-debug-toolbar==4.4.2
django-environ==0.11.2
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
